+++ static/js/components/output/executionTime.js
@@ -1,0 +1,3 @@
+import { getElement } from '../../utils/dom.js';
+
+export const execTime = getElement('exec-time');