+++ static/js/components/output/highlight.js
@@ -1,0 +1,5 @@
+// static/js/components/output/highlight.js
+import { getElement } from '../../utils/dom.js';
+
+export const highlightText = getElement('highlight-text');
+export const highlightRegex = getElement('highlight-regex');