+++ static/js/components/output/webTime.js
@@ -1,0 +1,4 @@
+// static/js/components/output/webTime.js
+import { getElement } from '../../utils/dom.js';
+
+export const webTime = getElement('web-time');