+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-01-21
+
+### Added
+
+- link to official website "https://openregex.com" in web interface
+- regex output and highlight connection in web interface
+- regex CheatSheet in web interface
+- debug information in web interface
+
+### Changed
+
+- reorganize code for `*.js` and `*.css` files
+- reorder colors group in web interface
+- black theme for web interface
+
+### Fixed
+
+- error in web-browser console for `*.js` files
 
 ## [1.0.0] - 2025-01-10